requirements/prod.txt
Django==4.2.9
django-betterforms==2.0.0
django-cleanup==8.1.0
django-tinymce==3.7.1
pillow==10.2.0
python-dotenv==1.0.1
python-slugify==8.0.4
sorl-thumbnail==12.10.0